Removed standalone typed getters
They were only used in SettingsType and tests, better to test what we actually use in the app.
